Added confirmation page decorate next

+++ src/src/main/java/pizzasystem/CartItem.java
@@ -6,9 +6,9 @@ import java.sql.SQLException;
 public class CartItem           //temporary class for testing remove this class with the proper one
 {
     private String item;
-    private float price;
+    private double price;
     private int quantity;
-    public CartItem(String item, float price, int quantity) {
+    public CartItem(String item, double price, int quantity) {
         this.item = item;
         this.price = price;
         this.quantity = quantity;
@@ -18,7 +18,7 @@ public class CartItem           //temporary class for testing remove this class 
         return item;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
